now fully grabbing market card from comps for simplified.
